LayoutTests/imported/w3c/web-platform-tests/webmessaging/broadcastchannel/sandbox-expected.txt



FAIL Creating BroadcastChannel in an opaque origin assert_equals: expected "Created" but got "Exception: ReferenceError"
